requirements.txt
python-dotenv
azure-identity
azure-ai-projects
azure-ai-inference
openai
requests